refactor: create sales app in frontend
