+++ frontend/src/App.tsx
@@ -13,11 +13,11 @@ import Reports from './components/Reports';
 import Navbar from './components/Navbar';
 
 // New sales components
-import SalesList from '@/components/sales/SalesList';
-import SaleForm from './components/sales/SaleForm';
-import SaleDetail from './components/sales/SaleDetail';
-import QuickSale from './components/sales/QuickSale';
-import SalesDashboard from './components/sales/SalesDashboard';
+import SalesList from '@/apps/sales/components/SalesList';
+import SaleForm from './apps/sales/components/SaleForm';
+import SaleDetail from './apps/sales/components/SaleDetail';
+import QuickSale from './apps/sales/components/QuickSale';
+import SalesDashboard from './apps/sales/components/SalesDashboard';
 
 // New invoice components
 import InvoicesList from './components/invoices/InvoicesList';
