+++ frontend/src/apps/analytics/pages/SalesAnalyticsDashboard.tsx
@@ -168,7 +168,7 @@ export const SalesAnalyticsDashboard: React.FC = () => {
         </div>
 
         {/* Heatmap Area - Bottom Full Width */}
-        <div style={{ gridArea: 'heatmap' }} className="bg-white rounded-xl shadow-lg p-6 flex flex-col">
+        {/* <div style={{ gridArea: 'heatmap' }} className="bg-white rounded-xl shadow-lg p-6 flex flex-col">
           <div className="flex justify-between items-center mb-4">
             <h2 className="text-xl font-semibold text-gray-800">خريطة نشاط المبيعات</h2>
             <div className="flex space-x-2 space-x-reverse">
@@ -212,7 +212,7 @@ export const SalesAnalyticsDashboard: React.FC = () => {
               className="h-full"
             />
           </div>
-        </div>
+        </div> */}
 
         {/* Empty State Overlay */}
         {!timeSeriesLoading && !heatmapLoading &&
