+++ frontend/src/apps/sales/hooks/useSalesCalculations.ts
